move code to io.ts and revert package.json change
