+++ packages/deployer/package.json
@@ -26,8 +26,5 @@
     "serverless": "^2.8.0",
     "serverless-plugin-optimize": "^4.1.4-rc.1",
     "serverless-plugin-typescript": "^1.1.9"
-  },
-  "devDependencies": {
-    "@types/node": "^14.14.22"
   }
 }
